Add SQLite database
- add sqlite database and wordDAO
- implement swap, add, remove function
- add sample data as first install

+++ app/src/main/java/yuntech/b10517012/visualchat/ui/IEditWord.kt
@@ -1,5 +1,5 @@
 package yuntech.b10517012.visualchat.ui
 
 interface IEditWord {
-    fun alertEditWord(index: Int, word: String)
+    fun alertEditWord(index: Long, word: String)
 }
